+++ games_services_platform_interface/lib/src/game_services_platform_impl.dart
@@ -186,11 +186,13 @@ class MethodChannelGamesServices extends GamesServicesPlatform {
   Future<String?> getAuthCode(
     String clientID, {
     bool forceRefreshToken = false,
+    List<String> additionalScopes = const [],
   }) =>
       Device.isPlatformAndroid
           ? _methodChannel.invokeMethod("getAuthCode", {
               "clientID": clientID,
               "forceRefreshToken": forceRefreshToken,
+              "additionalScopes": additionalScopes,
             })
           : Future.value(null);